gui/requirements.txt
PyQt5
pyserial
pyqtgraph